@@ -20,17 +20,17 @@
   ],
   "require": {
     "php": ">=8.1",
-    "symfony/http-kernel": "^6.3",
+    "symfony/http-kernel": "^6.4",
     "doctrine/orm": "^2.14",
-    "symfony/config": "^6.3",
-    "symfony/dependency-injection": "^6.3",
+    "symfony/config": "^6.4",
+    "symfony/dependency-injection": "^6.4",
     "doctrine/collections": "^2.1",
-    "symfony/event-dispatcher": "^6.3",
-    "symfony/expression-language": "^6.3",
-    "symfony/serializer": "^6.3",
-    "symfony/property-access": "^6.3",
+    "symfony/event-dispatcher": "^6.4",
+    "symfony/expression-language": "^6.4",
+    "symfony/serializer": "^6.4",
+    "symfony/property-access": "^6.4",
     "stof/doctrine-extensions-bundle": "^1.7",
-    "symfony/validator": "^6.3"
+    "symfony/validator": "^6.4"
   },
   "autoload": {
     "psr-4": {
