@@ -14,7 +14,7 @@ class NameNode extends Node
         );
     }
 
-    public function compile(Compiler $compiler)
+    public function compile(Compiler $compiler): void
     {
         $compiler->raw($this->attributes['name']);
     }
